+++ test/util/hashable_test.dart
@@ -1,0 +1,25 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import '../mocks/hashable_mock.dart';
+
+///
+///
+///
+void main() {
+  ///
+  group(
+    'Hashable',
+    () {
+      Map<HashableMock, int> domain = <HashableMock, int>{
+        HashableMock(): 44916520,
+      };
+
+      for (final MapEntry<HashableMock, int> input in domain.entries) {
+        test(
+          '${input.key} // ${input.value}',
+          () => expect(input.key.hashCode, input.value),
+        );
+      }
+    },
+  );
+}